requirements.txt
asgiref==3.8.1
attrs==25.1.0
click==8.1.8
Django==5.1.5
django-constance==4.3.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
uvicorn==0.34.0
